refactor: extract parsers from signal/client.rs into signal/parse.rs (stage A of #400)
Mechanical split along the existing seam. signal/client.rs was 3,899 LOC
mixing two concerns that share no state: lifecycle/RPC stubs (SignalClient
struct, child process management, 26 send_* methods) and JSON parsers
(parse_rpc_result and ~17 free parse_* functions over serde_json::Value).

Stage A moves all parsers plus their #[cfg(test)] mod tests block (42
test functions / 49 cases after rstest) into a new src/signal/parse.rs.
parse_rpc_result and parse_signal_event drop from pub fn to pub(super);
no callers outside the signal:: tree existed. The submodule is registered
as pub(crate) mod parse in signal/mod.rs.

- src/signal/client.rs: 3,899 -> 1,156 LOC
- src/signal/parse.rs: new, 2,758 LOC
- src/signal/mod.rs: +1 line
- 549 tests still passing, clippy clean, fmt clean
- App field-count baseline 66/66 unchanged

Stage B (send_rpc helper to dedup the 26 send_* methods) lands in a
follow-up commit on this same branch.

Refs #400
