+++ src/signal/mod.rs
@@ -1,5 +1,6 @@
-//! signal-cli integration: child process bridge ([`client`]) and wire types
-//! ([`types`]).
+//! signal-cli integration: child process bridge ([`client`]), JSON-RPC frame
+//! parsers ([`parse`]), and wire types ([`types`]).
 
 pub mod client;
+pub(crate) mod parse;
 pub mod types;
